% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krw_mafy_add_abundantieklasse.R
\name{add_abundantieklasse}
\alias{add_abundantieklasse}
\title{Toevoegen van de abundantieklasse}
\usage{
add_abundantieklasse(df, bedekking, type = "percentage")
}
\arguments{
\item{df}{Een dataframe met vegetatiegegevens}

\item{bedekking}{Naam van de kolom met de waarden van de bedekking. Mag met en zonder "".}

\item{type}{\code{"percentage"} of \code{"klasse"}. Geeft aan of in de kolom \code{bedekking} percentages of
klassen (1 t/m 9) staan.}
}
\value{
Het input dataframe met een toegevoegde kolom \code{abundantieklasse}
}
\description{
Deze functie voegt aan een dataframe een kolom \code{abundantieklasse} toe met de driedelige
abundantieklassen die bij de KRW-toetsing van macrofyten wordt gebruikt. De functie kan
zowel percentages als de 9-delige STOWA-klasse (afgeleid van Tansley en/of Braun-Blanquet)
omzetten in de abundantieklassen.
}
\note{
Missende en/of ongeldige waarden (bijv. 0, <0% klasse 1.5 of klasse 10) worden
omgezet in \code{NA}. In deze gevallen geeft de functie wel een melding dat deze
waarden geen correcte waarden zijn.
}
\examples{
test <- tibble::tribble(
~naam, ~bedekkingspercentage, ~klasse,
"c",                     1,       1,
"e",                     5,       3,
"f",                    25,       4,
"g",                    49,       7,
"h",                    50,       8,
"i",                    51,       9
)

add_abundantieklasse(test, bedekkingspercentage, type = "percentage")
add_abundantieklasse(test, "bedekkingspercentage", type = "percentage")

add_abundantieklasse(test, klasse, type = "klasse")

}
