% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data_description.R
\docType{data}
\name{ept_families}
\alias{ept_families}
\title{EPT-Families}
\format{
Vector met EPT-families
}
\source{
\href{https://www.aquo-kit.nl/}{Aquo-kit}
}
\usage{
ept_families
}
\description{
Families die behoren tot de groep Ephemeroptera (haften), Plecoptera (steenvliegen) en Trichoptera (kokerjuffers).
}
\keyword{datasets}
