% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data_description.R
\docType{data}
\name{ept_taxa}
\alias{ept_taxa}
\title{EPT-taxa}
\format{Vector met EPT-taxa}
\source{
\href{https://www.aquo-kit.nl/}{Aquo-kit}
}
\usage{
ept_taxa
}
\description{
Taxa die behoren tot de groep Ephemeroptera (haften), Plecoptera (steenvliegen) en Trichoptera (kokerjuffers).
}
\keyword{datasets}
