% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_description.R
\docType{package}
\name{krw}
\alias{krw}
\alias{krw-package}
\title{krw: Een package om EKR-scores uit te rekenen}
\description{
Functies voor om EKR-scores te bepalen volgens de maatlatten van 2018.
Deze functies zijn vooralsnog uitsluitend bedoeld voor de toetsing van M-typen.
Zie voor meer informatie: \url{https://github.com/RedTent/krw}
}
