% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data_description.R
\docType{data}
\name{krw_mafa_categorie_2018}
\alias{krw_mafa_categorie_2018}
\title{KRW-maatlatten: Macrofauna categorieen}
\format{
Dataframe
\itemize{
\item \code{biotaxon.naam} Naam van het biotaxon, inclusief ondersoorten die niet expliciet in het
maatlatdocument zijn opgenomen.
\item \code{krwwatertype.code} KRW-watertypecode
\item \code{categorie} Categorie van het taxon voor het betreffende watertype.
Mogelijk categorieen zijn:
\itemize{
\item \strong{P} - Positief taxon
\item \strong{K} - Kenmerkend taxon
\item \strong{N} - Negatief taxon
}
}
}
\source{
\href{https://www.aquo-kit.nl/}{Aquo-kit}
}
\usage{
krw_mafa_categorie_2018
}
\description{
Aanduiding voor macrofauna taxa of een taxon voor een bepaald watertype behoort
tot de positieve (P), kenmerkende (K) of negatieve (N) soorten. Deze tabel gaat over alle watertypen
maar alleen over de maatlatten van 2018.
}
\details{
Macrofauna kan getoetst worden voor de Kaderrichtlijn Water d.m.v. zogenaamde maatlatten.
De documentatie voor deze maatlatten kan gevonden worden op de website van \href{www.stowa.nl}{STOWA}.
Het betreft:
\itemize{
\item Referenties en maatlatten voor natuurlijke watertypen voor de Kaderrichtlijn Water 2021-2027
\item Omschrijving MEP en maatlatten voor sloten en kanalen voor de Kaderrichtlijn Water 2021-2027
}
}
\note{
In deze tabel zijn alleen de categorieën van de maatlat van 2018 opgenomen.
}
\section{Metadata dataset}{

\itemize{
\item \strong{Naam:} Macrofauna-categorieen
\item \strong{Oorspronkelijke Naam:} Somparametersamenstelling, de macrofauna-categorieen zijn hier een subset van.
\item \strong{Bronhouder:} IHW (voor Aquo-kit) en STOWA voor de onderliggende data
\item \strong{Herkomst:} Aquo-kit \url{https://www.aquo-kit.nl/}
\item \strong{Auteurs:} Diverse
\item \strong{Datum dataset:} 2018
\item \strong{Datum opname dataset:} 6-7-2019
\item \strong{Gebruiksrechten:} Onbeperkt
\item \strong{Soortgroep:} Macrofauna
}
}

\section{Bewerkingen}{


De kolomnamen zijn omgezet naar lowercase.
}

\examples{

head(krw_mafa_categorie_2018)
}
\seealso{
AANVULLEN
}
\keyword{datasets}
