% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data_description.R
\docType{data}
\name{krw_mafa_constanten_2018}
\alias{krw_mafa_constanten_2018}
\title{KRW-maatlatten: Macrofauna constanten}
\format{Dataframe
\itemize{
\item \code{krwwatertype.code} KRW-watertypecode
\item \code{chloride_min} Ondergrens van de chlorideconcentratie in mg/l. Alleen voor watertype M30.
\item \code{chloride_max}Bovengrens van de chlorideconcentratie in mg/l. Alleen voor watertype M30.
\item \code{kmmax} Maximaal aantal kenmerkende taxa. Alleen voor natuurlijke watertypen
\item \code{ptmax} Maximaal aantal positieve taxa. Alleen voor kunstmatige watertypen
\item \code{dnmax} Maximaal aandeel in de abundantie. Alleen voor
}}
\source{
\href{https://www.aquo-kit.nl/}{Aquo-kit}
}
\usage{
krw_mafa_constanten_2018
}
\description{
Constanten per watertype die nodig zijn voor de berekening van de EKR voor macrofauna. In deze tabel
zijn alleen de constanten voor de maatlat van 2018 opgenomen.
}
\details{
Macrofauna kan getoetst worden voor de Kaderrichtlijn Water d.m.v. zogenaamde maatlatten.
De documentatie voor deze maatlatten kan gevonden worden op de website van \href{www.stowa.nl}{STOWA}.
Het betreft:
\itemize{
\item Referenties en maatlatten voor natuurlijke watertypen voor de Kaderrichtlijn Water 2021-2027
\item Omschrijving MEP en maatlatten voor sloten en kanalen voor de Kaderrichtlijn Water 2021-2027
}
}
\note{
In deze tabel zijn alleen de constanten van de maatlat van 2018 opgenomen.
}
\section{Metadata dataset}{

\itemize{
\item \strong{Naam:} Macrofauna-constanten
\item \strong{Oorspronkelijke Naam:} 35xMacrofauna-constanten-berekening-kwaliteitselement_yyyymmdduummss.csv
\item \strong{Bronhouder:} IHW (voor Aquo-kit) en STOWA voor de onderliggende data
\item \strong{Herkomst:} Aquo-kit \url{https://www.aquo-kit.nl/}
\item \strong{Auteurs:} Diverse
\item \strong{Datum dataset:} 2018
\item \strong{Datum opname dataset:} 6-7-2019
\item \strong{Gebruiksrechten:} Onbeperkt
\item \strong{Soortgroep:} Macrofauna
}
}

\section{Bewerkingen}{


De kolomnamen zijn omgezet naar lowercase.

De kolomnamen van de chlorideconcentraties zijn omgezet naar een meer computervriendelijke naam.
}

\examples{

head(krw_mafa_constanten_2018)
}
\seealso{
AANVULLEN
}
\keyword{datasets}
