% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krw_mafa_toetsing.R
\name{krw_mafa_ekr}
\alias{krw_mafa_ekr}
\title{KRW-toetsing macrofauna}
\usage{
krw_mafa_ekr(df, krwwatertype.code, biotaxon.naam, waarde,
  verbose = FALSE)
}
\arguments{
\item{df}{Een dataframe met macrofauna-data. Als het dataframe meerdere monsters bevat
moet dit dataframe een \emph{grouped dataframe} worden met de functie \code{\link[dplyr:group_by]{dplyr::group_by()}}}

\item{krwwatertype.code}{Naam van de kolom met de KRW-watertype code}

\item{biotaxon.naam}{Naam van de kolom met de taxonnamen.}

\item{waarde}{Naam van de kolom met aantallen van het betreffende taxon.}

\item{verbose}{Logical. Als \code{TRUE} dan worden behalve de EKR ook de onderliggende parameters
en constanten in de uitvoer meegenomen. Deze zijn vooral relevant om e.e.a. te kunnen controleren.}
}
\value{
Een dataframe met groepkolommen (alleen indien van toepassing), het watertype en de macrofauna EKR.
}
\description{
Deze functie voert de KRW-toetsing voor macrofauna uit voor een of meerdere monsters. De toetsing is gebaseerd op de
maatlatten uit het jaar 2018.
}
\details{
De kolomnamen van de kolommen met taxon, watertype en waarde kunnen zowel met als zonder quotes ("")
worden ingevoerd door het gebruik van non-standard-evaluation (NSE).
}
\note{
Het is van belang dat de biotaxonnamen in de TWN-lijst voorkomen. De KRW-watertypecode dient overeen te komen met de
codes uit de codes uit de aquo-standaard.

Deze functie is niet geschikt voor de watertypen M30 en R8.
}
\examples{
\dontrun{

# Een enkel monster

krw_mafa_ekr(df_1_monster, watertype, taxon, aantallen)
krw_mafa_ekr(df_1_monster, "watertype", "taxon", "aantallen")


# Meerdere monsters in 1 dataframe

grouped_df <- group_by(df_meerdere_monsters, meetpunt, datum)
krw_mafa_ekr(grouped_df, watertype, taxon, aantallen)  

# of met het pipe-teken (\%>\%)

df_meerdere_monsters \%>\% 
  group_by(meetpunt, datum) \%>\% 
  krw_mafa_ekr(watertype, taxon, aantallen)

}


}
