% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krw_mafy_toetsing.R
\name{krw_mafy_ekr_ss}
\alias{krw_mafy_ekr_ss}
\title{KRW-toetsing macrofyten soortensamenstelling}
\usage{
krw_mafy_ekr_ss(df, krwwatertype.code, biotaxon.naam, abundantieklasse,
  verbose = FALSE)
}
\arguments{
\item{df}{Een dataframe met macrofyten. Als het dataframe meerdere monsters bevat
moet dit dataframe een \emph{grouped dataframe} worden met de functie \code{\link[dplyr:group_by]{dplyr::group_by()}}}

\item{krwwatertype.code}{Naam van de kolom met de KRW-watertype code}

\item{biotaxon.naam}{Naam van de kolom met de taxonnamen.}

\item{abundantieklasse}{Naam van de kolom met de abundantieklassen van het betreffende taxon.
De abundantieklasse dient de waarde 1, 2, of 3 te bevatten. De functie AANVULLEN kan worden gebruikt
om om te rekenen naar abundantieklassen.}

\item{verbose}{Logical. Als \code{TRUE} dan worden behalve de EKR ook de onderliggende parameters
en constanten in de uitvoer meegenomen. Deze zijn vooral relevant om e.e.a. te kunnen controleren.}
}
\value{
Een dataframe met groepkolommen (alleen indien van toepassing), het watertype en:
\itemize{
\item \code{ekr_ss_mafy} - de EKR voor de soortensamenstelling
\item \code{ekr_ss_hydrofyten} - de deel-EKR voor de soortensamenstelling hydrofyten (alleen sloten en kanalen)
\item \code{ekr_ss_helofyten} - de deel-EKR voor de soortensamenstelling helofyten (alleen sloten en kanalen)
}
}
\description{
Deze functie voert de KRW-toetsing uit voor de macrofyten-soortensamenstelling voor een of meerdere monsters.
De toetsing is gebaseerd op de maatlatten uit het jaar 2018.
}
\details{
De kolomnamen van de kolommen met taxon, watertype en waarde kunnen zowel met als zonder quotes ("")
worden ingevoerd door het gebruik van non-standard-evaluation (NSE).
}
\note{
Het is van belang dat de biotaxonnamen in de TWN-lijst voorkomen. De KRW-watertypecode dient overeen te komen met de
codes uit de codes uit de aquo-standaard.
}
\examples{
\dontrun{

# Een enkel monster

krw_mafy_ekr_ss(df_1_monster, watertype, taxon, abundantie)
krw_mafy_ekr_ss(df_1_monster, "watertype", "taxon", "abundantie")

 
# Meerdere monsters in 1 dataframe

grouped_df <- group_by(df_meerdere_monsters, meetpunt, datum)
krw_mafy_ekr_ss(grouped_df, watertype, taxon, abundantie)  

# of met het pipe-teken (\%>\%)

df_meerdere_monsters \%>\% 
  group_by(meetpunt, datum) \%>\% 
  krw_mafy_ekr_ss(watertype, taxon, abundantie)

}
}
