% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data_description.R
\docType{data}
\name{krw_mafy_taxonscore_2018}
\alias{krw_mafy_taxonscore_2018}
\title{KRW-maatlatten: Macrofytenscores}
\format{Dataframe
\itemize{
\item \code{groep} Hydrofyten of Helofyten. Alleen voor sloten en kanalen, M1 t/m M10.
\item \code{biotaxon.naam} Naam van het biotaxon, inclusief ondersoorten die niet expliciet in het
maatlatdocument zijn opgenomen.
\item \code{krwwatertype.code} KRW-watertypecode
\item \code{abundantieklasse} Abundantieklasse: 1, 2 of 3
\item \code{score} De score voor de betreffende taxon + watertype + abundantie
}}
\source{
\href{https://www.aquo-kit.nl/}{Aquo-kit}
}
\usage{
krw_mafy_taxonscore_2018
}
\description{
Score per watertype, taxon en abundantie. Dit is een gecombineerde tabel gebaseerde op de
macrofytencategorieen en de macrofytenscores. Alleen de scores voor de maatlatten van
2018 zijn opgenomen.
}
\details{
Macrofyten kunnen getoetst worden voor de Kaderrichtlijn Water d.m.v. zogenaamde maatlatten.
De documentatie voor deze maatlatten kan gevonden worden op de website van \href{www.stowa.nl}{STOWA}.
Het betreft:
\itemize{
\item Referenties en maatlatten voor natuurlijke watertypen voor de Kaderrichtlijn Water 2021-2027
\item Omschrijving MEP en maatlatten voor sloten en kanalen voor de Kaderrichtlijn Water 2021-2027
}
}
\note{
In deze tabel zijn alleen de scores van de maatlat van 2018 opgenomen.
}
\section{Metadata dataset}{

\itemize{
\item \strong{Naam:} Macrofyten taxonscores 2018
\item \strong{Oorspronkelijke Namen:}
\itemize{
\item 250aMacrofyten-categorieen_yyyymmdduummss
\item 250cMacrofyten_scores_yyyymmdduummss
}
\item \strong{Bronhouder:} IHW (voor Aquo-kit) en STOWA voor de onderliggende data
\item \strong{Herkomst:} Aquo-kit \url{https://www.aquo-kit.nl/}
\item \strong{Auteurs:} Diverse
\item \strong{Datum dataset:} 2018
\item \strong{Datum opname dataset:} 25-6-2019
\item \strong{Gebruiksrechten:} Onbeperkt
\item \strong{Soortgroep:} Macrofyten
}
}

\section{Bewerkingen}{


De twee originele tabellen zijn gecombineerd tot 1 tabel.

De kolomnamen zijn omgezet naar lowercase.
}

\examples{

head(krw_mafy_taxonscore_2018)
}
\seealso{
AANVULLEN
}
\keyword{datasets}
