% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_description.R
\docType{package}
\name{krw}
\alias{krw}
\title{krw: Een package om EKR-scores uit te rekenen}
\description{
Functies voor om EKR-scores te bepalen volgens de maatlatten van 2018.
Zie voor meer informatie de package-website \url{https://redtent.github.io/krw/}
of de Github pagina: \url{https://github.com/RedTent/krw}
}
